@@ -1,37 +1,41 @@
-const colors = require('tailwindcss/colors');
-module.exports = {
-  content: [
-    "./src/**/*.{js,jsx,ts,tsx}",
-  ],
+/** @type {import('tailwindcss').Config} */
+export default {
+  content: ["./index.html", "./src/**/*.{js,ts,jsx,tsx}"],
+  darkMode: "class",
   theme: {
-    screens: {
-      sm: '480px',
-      md: '1197px',
-      lg: '1485px',
-      xl: '1800px',
-    },
-    letterSpacing: {
-      wide:'1rem',
-    },
     extend: {
       colors: {
-        'cyan': colors.cyan,
-        'teal':colors.teal,
+        primary: {
+          50: "#f0f9ff",
+          100: "#e0f2fe",
+          200: "#bae6fd",
+          300: "#7dd3fc",
+          400: "#38bdf8",
+          500: "#0ea5e9",
+          600: "#0284c7",
+          700: "#0369a1",
+          800: "#075985",
+          900: "#0c4a6e",
+        },
       },
-      zIndex: {
-        '100': '100',
-        '1000':'1000'
-      },
-      backgroundImage: {
-        "pattern":" url('/src/asset/dust_texture.svg')"
+      fontFamily: {
+        sans: ["Inter", "system-ui", "sans-serif"],
       },
       animation: {
-        
+        "fade-in": "fadeIn 0.6s ease-in",
+        "slide-up": "slideUp 0.6s ease-out",
       },
       keyframes: {
-        
-      }
+        fadeIn: {
+          "0%": { opacity: "0" },
+          "100%": { opacity: "1" },
+        },
+        slideUp: {
+          "0%": { transform: "translateY(20px)", opacity: "0" },
+          "100%": { transform: "translateY(0)", opacity: "1" },
+        },
+      },
     },
   },
   plugins: [],
-}
+};
